+++ python/tests/test_phylo_formats.py
@@ -89,22 +89,73 @@ class TestNewick(TreeExamples):
     external Newick parser.
     """
 
-    def verify_newick_topology(self, tree, root=None, node_labels=None):
+    # 3 methods to return example tree sequences with internal samples:
+    # (copied from test_highlevel.py)
+    def all_nodes_samples_example(self):
+        n = 5
+        ts = msprime.simulate(n, random_seed=10, mutation_rate=5)
+        assert ts.num_mutations > 0
+        tables = ts.dump_tables()
+        nodes = tables.nodes
+        flags = nodes.flags
+        # Set all nodes to be samples.
+        flags[:] = tskit.NODE_IS_SAMPLE
+        nodes.flags = flags
+        return tables.tree_sequence()
+
+    def only_internal_samples_example(self):
+        n = 5
+        ts = msprime.simulate(n, random_seed=10, mutation_rate=5)
+        assert ts.num_mutations > 0
+        tables = ts.dump_tables()
+        nodes = tables.nodes
+        flags = nodes.flags
+        # Set just internal nodes to be samples.
+        flags[:] = 0
+        flags[n:] = tskit.NODE_IS_SAMPLE
+        nodes.flags = flags
+        return tables.tree_sequence()
+
+    def mixed_node_samples_example(self):
+        n = 5
+        ts = msprime.simulate(n, random_seed=10, mutation_rate=5)
+        assert ts.num_mutations > 0
+        tables = ts.dump_tables()
+        nodes = tables.nodes
+        flags = nodes.flags
+        # Set a mixture of internal and leaf samples.
+        flags[:] = 0
+        flags[n // 2 : n + n // 2] = tskit.NODE_IS_SAMPLE
+        nodes.flags = flags
+        return tables.tree_sequence()
+
+    def verify_newick_topology(
+        self, tree, root=None, node_labels=None, include_branch_lengths=True
+    ):
         if root is None:
             root = tree.root
-        ns = tree.newick(precision=16, root=root, node_labels=node_labels)
+        ns = tree.newick(
+            precision=16,
+            root=root,
+            node_labels=node_labels,
+            include_branch_lengths=include_branch_lengths,
+        )
         if node_labels is None:
             leaf_labels = {u: str(u + 1) for u in tree.leaves(root)}
         else:
             leaf_labels = {u: node_labels[u] for u in tree.leaves(root)}
-        newick_tree = newick.loads(ns)[0]
+        # default newick lib outputs 0.0 if length is None => replace the length_parser
+        newick_tree = newick.loads(
+            ns, length_parser=lambda x: None if x is None else float(x)
+        )[0]
         leaf_names = newick_tree.get_leaf_names()
         assert sorted(leaf_names) == sorted(leaf_labels.values())
         for u in tree.leaves(root):
             name = leaf_labels[u]
             node = newick_tree.get_node(name)
             while u != root:
-                self.assertAlmostEqual(node.length, tree.branch_length(u))
+                branch_len = tree.branch_length(u) if include_branch_lengths else None
+                self.assertAlmostEqual(node.length, branch_len)
                 node = node.ancestor
                 u = tree.parent(u)
             assert node.ancestor is None
@@ -161,6 +212,34 @@ class TestNewick(TreeExamples):
         assert [n.name for n in root.walk()] == [labels[tree.root]] + [
             None for _ in range(len(list(tree.nodes())) - 1)
         ]
+
+    def test_no_lengths(self):
+        t = msprime.simulate(5, random_seed=2).first()
+        self.verify_newick_topology(t, include_branch_lengths=False)
+
+    def test_samples_differ_from_leaves(self):
+        for ts in (
+            self.all_nodes_samples_example(),
+            self.only_internal_samples_example(),
+            self.mixed_node_samples_example(),
+        ):
+            for t in ts.trees():
+                self.verify_newick_topology(t)
+
+    def test_no_lengths_equiv(self):
+        for ts in (
+            self.all_nodes_samples_example(),
+            self.only_internal_samples_example(),
+            self.mixed_node_samples_example(),
+        ):
+            for t in ts.trees():
+                newick_nolengths = t.newick(include_branch_lengths=False)
+                newick_nolengths = newick.loads(newick_nolengths)[0]
+                newick_lengths = t.newick()
+                newick_lengths = newick.loads(newick_lengths)[0]
+                for node in newick_lengths.walk():
+                    node.length = None
+                assert newick.dumps(newick_nolengths) == newick.dumps(newick_lengths)
 
 
 class TestNexus(TreeExamples):